+++ test/widget/user_profile_test.dart
@@ -506,7 +506,7 @@ void main() async {
 
     expect(find.text('user name'), findsAny);
     expect(find.byKey(const Key('Present')), findsOneWidget);
-    await tester.dragUntilVisible(find.byKey(const Key('UserNum')),
+    await tester.dragUntilVisible(find.byKey(const Key('NumCopyable')),
         find.byKey(const Key('UserScrollable')), const Offset(1, 1));
     await tester.pumpAndSettle(const Duration(seconds: 2));
     expect(find.text('5769space2360space9862space1822'), findsOneWidget);